refactor: remove unnecessarily long paths in imports and mocks
